some restructuring and entrypoints
