+++ pyproject.toml
@@ -35,3 +35,7 @@ dev = ["pytest>=6.0", "psutil", "hidefix", "h5py", "netCDF4", "zarr", "tensorsto
 
 [project.entry-points."xarray.backends"]
 om = "omfiles.xarray_backend:OmXarrayEntrypoint"
+
+[project.entry-points."zarr.codecs"]
+"omfiles.pfor" = "omfiles.omfiles_zarr_codecs:PforCodec"
+"omfiles.pfor_serializer" = "omfiles.omfiles_zarr_codecs:PforSerializer"
