+++ python/omfiles/__init__.py
@@ -1,4 +1,4 @@
-from . import types, xarray_backend, omfiles_numcodecs
+from . import omfiles_zarr_codecs, types, xarray_backend
 from .omfiles import OmFilePyReader, OmFilePyWriter, OmVariable
 
-__all__ = ["OmFilePyReader", "OmFilePyWriter", "OmVariable", "omfiles_numcodecs", "xarray_backend", "types"]
+__all__ = ["OmFilePyReader", "OmFilePyWriter", "OmVariable", "omfiles_zarr_codecs", "xarray_backend", "types"]
